+++ plone/dexterity/tests/test_schema.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from pkg_resources import get_distribution
 from plone.dexterity import schema
 from plone.dexterity.fti import DexterityFTI
 from plone.dexterity.interfaces import IContentType
@@ -15,9 +14,6 @@ from zope.interface.interface import InterfaceClass
 
 import unittest
 import zope.schema
-
-
-has_zope4 = get_distribution('Zope2').version.startswith('4')
 
 
 class TestSchemaModuleFactory(MockTestCase):
@@ -164,7 +160,6 @@ class TestSchemaModuleFactory(MockTestCase):
             schema.portalTypeToSchemaName('type one.two', '', 'prefix')
         )
 
-    @unittest.skipIf(has_zope4, 'Broken with zope4, see https://community.plone.org/t/problems-with-mocktestcase-in-plone-dexterity/1484')  # noqa
     def test_portalTypeToSchemaName_looks_up_portal_for_prefix(self):
         portal_mock = self.mocker.mock()
         self.expect(