+++ plone/dexterity/tests/test_utils.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from pkg_resources import get_distribution
 from plone.dexterity import utils
 from plone.dexterity.fti import DexterityFTI
 from plone.mocktestcase import MockTestCase
@@ -7,12 +6,8 @@ from plone.mocktestcase import MockTestCase
 import unittest
 
 
-has_zope4 = get_distribution('Zope2').version.startswith('4')
-
-
 class TestUtils(MockTestCase):
 
-    @unittest.skipIf(has_zope4, 'Broken with zope4, see https://community.plone.org/t/problems-with-mocktestcase-in-plone-dexterity/1484')  # noqa
     def test_getAdditionalSchemata(self):
         from plone.dexterity.interfaces import IDexterityFTI
         from plone.behavior.interfaces import IBehavior